+++ ToArcGISEarth/ToArcGISEarth/ToArcGISEarth.csproj
@@ -104,12 +104,15 @@
     <AddInContent Include="DarkImages\AddInDesktop32.png" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AddLayerCheckBox.cs" />
     <Compile Include="AutomationAPI\EarthNamedpipeAPIUtils.cs" />
     <Compile Include="AutomationAPI\IEarthNamedpipeService.cs" />
     <Compile Include="Module1.cs" />
     <Compile Include="AutomationAPI\ProcessUtils.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ViewLinkEarthButton.cs" />
+    <Compile Include="RemoveLayerCheckBox.cs" />
+    <Compile Include="ViewLinkCheckBox.cs" />
+    <Compile Include="ConnectToArcGISEarthButton.cs" />
   </ItemGroup>
   <ItemGroup>
     <AddInContent Include="DarkImages\GenericButtonBlue16.png" />
